"git checkout" via grep rather than full branch name
Hopping around feature branches can be a real pain, especially when following
conventions such as prefixing branch names with ticket numbers. This allows you
to checkout a branch via a grep phrase rather than a full branch name. If the
phrase matches zero or multiple branches, nothing is checked-out and a warning
is printed.